test: Move testcases for autodoc configurations to test_ext_autodoc_configs
